// src/version2/models/timeTrackingConfiguration.ts
/** Details of the time tracking configuration. */
export interface TimeTrackingConfiguration {
  /** The default unit of time applied to logged time. */
  defaultUnit: string;
  /** The format that will appear on an issue's _Time Spent_ field. */
  timeFormat: string;
  /** The number of days in a working week. */
  workingDaysPerWeek: number;
  /** The number of hours in a working day. */
  workingHoursPerDay: number;
}